Extract delete graph to own structure
